move function to lib
