+++ services/processor/processor/default_handlers/action_sync_users.py
@@ -7,64 +7,10 @@ from ftrack_common import (
     ServerAction,
     get_service_ftrack_icon_url,
 )
+from processor.lib import map_ftrack_users_to_ayon_users
 
 if typing.TYPE_CHECKING:
-    from ftrack_api.entity.user import User as FtrackUser
     from ftrack_api.entity.base import Entity as FtrackEntity
-
-
-def map_ftrack_users_to_ayon_users(
-    ftrack_users: List["FtrackUser"],
-    ayon_users: List[Dict[str, Any]],
-) -> Dict[str, Union[str, None]]:
-    """Map ftrack users to AYON users.
-
-    Mapping is based on 2 possible keys, email and username where email has
-    higher priority. Once AYON user is mapped it cannot be mapped again to
-    different user.
-
-    Args:
-        ftrack_users (List[ftrack_api.entity.user.User]): List of ftrack users.
-        ayon_users (List[Dict[str, Any]]): List of AYON users.
-
-    Returns:
-        Dict[str, Union[str, None]]: Mapping of ftrack user id
-            to AYON username.
-
-    """
-    mapping: Dict[str, Union[str, None]] = {
-        user["id"]: None
-        for user in ftrack_users
-    }
-    ayon_users_by_email: Dict[str, str] = {}
-    ayon_users_by_name: Dict[str, str] = {}
-    for ayon_user in ayon_users:
-        ayon_name = ayon_user["name"]
-        ayon_email = ayon_user["attrib"]["email"]
-        ayon_users_by_name[ayon_name.lower()] = ayon_name
-        if ayon_email:
-            ayon_users_by_email[ayon_email.lower()] = ayon_name
-
-    mapped_ayon_users: Set[str] = set()
-    for ftrack_user in ftrack_users:
-        ftrack_id: str = ftrack_user["id"]
-        ftrack_name: str = ftrack_user["username"]
-        ftrack_email: str = ftrack_user["email"]
-
-        if ftrack_email and ftrack_email.lower() in ayon_users_by_email:
-            ayon_name: str = ayon_users_by_email[ftrack_email.lower()]
-            if ayon_name not in mapped_ayon_users:
-                mapping[ftrack_id] = ayon_name
-                mapped_ayon_users.add(ayon_name)
-            continue
-
-        if ftrack_name in ayon_users_by_name:
-            ayon_name: str = ayon_users_by_name[ftrack_name]
-            if ayon_name not in mapped_ayon_users:
-                mapped_ayon_users.add(ayon_name)
-                mapping[ftrack_id] = ayon_name
-
-    return mapping
 
 
 class SyncUsersFromFtrackAction(ServerAction):
@@ -165,8 +111,8 @@ class SyncUsersFromFtrackAction(ServerAction):
         ftrack_projects: List[FtrackEntity] = session.query(
             "select id, full_name, is_private from Project"
         )
-        users_mapping: Dict[str, str] = map_ftrack_users_to_ayon_users(
-            ftrack_users, ayon_users
+        users_mapping: Dict[str, Union[str, None]] = (
+            map_ftrack_users_to_ayon_users(ftrack_users, ayon_users)
         )
         for ftrack_id, ayon_user_name in users_mapping.items():
             ftrack_user = ftrack_users_by_id[ftrack_id]
