+++ services/processor/processor/lib/__init__.py
@@ -1,8 +1,13 @@
+from .users import (
+    map_ftrack_users_to_ayon_users,
+)
 from .sync_from_ftrack import (
     SyncFromFtrack,
 )
 
 
 __all__ = (
+    "map_ftrack_users_to_ayon_users",
+
     "SyncFromFtrack",
 )
